refactor(Model): rename namespace MHWItemBoxTracker.Config to MHWItemBoxTracker.Model
