+++ ItemBoxTracker/src/Plugin/Model/DeprecatedItemBoxTrackerConfig.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel;
 
-namespace MHWItemBoxTracker.Config {
+namespace MHWItemBoxTracker.Model {
   public class DeprecatedItemBoxTrackerConfig {
     [DisplayName("Items to Track")]
     [Description("You can find item IDs here: https://github.com/Ezekial711/MonsterHunterWorldModding/wiki/Item-IDs")]
