+++ ItemBoxTracker/src/Plugin/ViewModel/ItemBoxTrackerViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.ObjectModel;
-using MHWItemBoxTracker.Config;
+using MHWItemBoxTracker.Model;
 using MHWItemBoxTracker.Utils;
 
 namespace MHWItemBoxTracker.ViewModels {
